Practice/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-grappelli==3.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
lxml==4.9.2
multidict==6.0.4
openpyxl==3.1.2
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyodbc==4.0.39
pytz==2023.3
redis==4.5.4
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
